utils: Compact radix tree
The tree uses integral type as a search key. On each level the local index
is one byte from the key, respectively for 32-bit key we have 4 levels.
The tree uses 2 memory packing techniques -- prefix compaction and growing
node layouts.

The prefix compaction is used when a node has only one child. In this case
such a node is replaced in its parent with this only child and the child in
question keeps "prefix:length" pair on board, that's used to check if the
short-cut lookup took the correct path.

The growing node layouts makes the nodes be of one of 4 kinds -- tiny, small,
medium and large. The node starts with the tiny layout and when the number
of keys on it grows, it gets transformed into bigger layout. Symmetrically,
when the keys are removed, the layouts are switched into opposite direction.

Tiny and small layouts are called "tables". The node consists of two arrays,
one with node indices, the other one with pointers/values.

   0: idx_a : value_a
   1: idx_b : value_b
   ...
   N: idx_N : value_N

The table is NOT sorted by index column. To "find" a node index on such a
node one needs to find the index on the first array, then peek the value from
the 2nd one. The tiny layout uses table of size 4, and the scan is linear
check.  The small layout use table of size 16 and the scan is SS4-4 lookup.

Medium layout is called "map". The map is array of 256 indices and array of
N pointers/values. To find an index in this layout one gets an index from the
first array and uses it as an index in the 2nd (provided it's < N).

  [ idx_0, idx_1 ... idx_255 ]
  [ value_0, value_1 ... value_N ]

In current implementation N is 48.

The large layout, called "array", is just an array of 256 pointers/values.

  [ value_0, ... value_256 ]

With this the tree memory footprint depends on the keys "sparseness". By the
sparseness measurement it's handy to use the value of

  ( maximum_key_value - minimum_key_value) / number_of_keys.

In case indices come sequentially, the sparseness is 1, in this case the
tree with 20 keys and more occupies 30% of the RB-tree with the same number
of keys.

When the sparseness value grows, the radix-tree grows as well. Only at the
sparseness value of ~100 the memory footpring becomes the same as for
RB-tree, then it grows firther. In its peak, the memory overhead is 2.5
times of the RB-tree.

For smaller trees (1 through 10 keys) the memory footprint is 0.9 through
1.5 times the RB-tree size.

The iterator is one-direction (for simplicity) but it enough for its main
target -- the sparse array of cells on a row. Also the iterator has an
.index() method that reports back the index of the entry at which it points.
This greatly simplifies the tree scans by the class row further.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>