row: Remove old storages
Now when the 3rd storage type (radix tree) is all in, old storage can
be safely removed.

The result is:

1. Memory footprint

sizeof(class row):  112 -> 16 bytes
sizeof(rows_entry): 232 -> 136 bytes

the "in cache" value depends on the number of cells.

     1 cell  :    744 -> 648 bytes    -12.9%
    10 cells :   1408 -> 1342 bytes    -4.6%
   100 cells :  11896 -> 10054 bytes  -15.4%
  1000 cells : 112696 -> 71174 bytes  -36.8%

Worth to note, that in its default run (5 cells) the footprint
is ~1% bigger, because 5 cells use radix tree node for 8 cells
and thus it's the biggest memory overhead around this size.

2. perf_mutation test is ~5% slower, because it uses 1 cell mutation
   and radix tree takes one more allocation to put the cell in, while
   current code just pushes the cell into the row's inline vector.
   It's the same loss as was the gain in f6bba394 when the vector
   storage type was introduced.

3. perf_simple_query average tps is ~1% less for reads, but it's
   within the measured deviation. For writes tps is ~3% less. It's
   less loss than the gain was when the vector was added.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>