client-state: Make options an enum set
Currently the client state keeps two bool-s on board -- is_internal
and is_thrift -- which are managed in a similar way. Soon there will
be the 3rd one, so to unify the management switch two bools into
a enum_set.

As a side effect -- give meaningful names to anonymous "true/false"-s

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>