vle: Optimized x86 version
Current uleb encoder does it's job in a loop. It can be done
better, but the betterness turned to differ significantly
depending on the CPU manufactiurer.

First, the optimized version does branch-less encoding just
by re-packing the given value with the help of classical
instructions (plus the clz one).

The second optimization can be done if using the bdep instruction
to replace half of the re-packing code, and this part is good on
Intel but is bad on AMD.

To compare the encoding routine itself I checked it on Intel i5,
i7 and Xeon and AMD threadripper chips.

Trivial optimization as compared to the loop version:

   Intel: 2.5 times faster
   AMD:   3.0 times faster

The same plus bdep instruction as compared to loop:

   Intel: 3.5 times faster
   AMD:   3.8 times slower (!)

The latter seems to be because the bdep is implemented in AMD's
microcode and slows things down.

So the bdep version is here, but is compiled out by default.

Next some heavier than the femto-benchmarks. All made on the
AMD Threadripper with non-bdep version.

* the added in this set LSA allocation test shows 15% smaller
allocation latency in dev mode and 40% smaller in release mode

* perf_collection test in release mode is 3% faster to fill the
B+tree with 5M keys. In dev mode the difference is within the
measured deviation.

* perf_mutation test doesn't show measurable difference

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>