cql3: Auto set initial_tablets if not specified
If the tablets feature is ON, but user didn't specify the
initial_tablets value when creating keyspace, tablets replication turns
on anyway with auto-calculated number of initial tables.

The latter is done based on current topology and RF so that all shards
are covered:

initial_tablets = max(nr_shards_in(dc) / RF_in(dc) for dc in datacenters)

The auto-set is performed in the prepare_options() helper which's
already in charge of auto-configuring the per-DC replication factor for
options that chose network replication strategy, but specify only the
'replication_factor' option, instead of per-DC values.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>