cql3: Introduce per-kyepsace tablets on/off toggle
The new option is called 'tablets' and can be either true or false. If
omitted, the code takes the tables feature state into account.

The option main purpose is to disable tablets replication when the feature
is enabled. When tablets are off and the option is true, create keyspace
is failed with configuration exception

Indentation is deliberately left broken in some places

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>